improv/dev-env: Update npm scripts to use yarn instead of npm
