+++ examples/standard.html
@@ -1,7 +1,7 @@
 <html>
 <head>
   <meta name="viewport" content="width=device-width">
-  <script type="module" src="./index.js"></script>
+  <script type="module" src="../dist/index.js"></script>
   <style>
     .examples {
       margin-top: 20px;
@@ -32,7 +32,7 @@
     </media-control-bar>
   </media-controller>
   <div class="examples">
-    <a href="/examples/">View more examples</a>
+    <a href="./">View more examples</a>
   </div>
 </body>
 </html>
