+++ package.json
@@ -10,8 +10,9 @@
   ],
   "scripts": {
     "clean": "rimraf dist",
-    "build": "npm run clean && node build.js",
-    "dev": "npm run clean && node build.js --dev",
+    "build": "yarn clean && webpack --config webpack.prod.js",
+    "dev": "yarn start",
+    "start": "webpack serve --config webpack.dev.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "publish-patch": "yarn run build && np patch --no-tests --any-branch"
   },
@@ -35,23 +36,12 @@
   },
   "homepage": "https://github.com/muxinc/media-chrome#readme",
   "devDependencies": {
-    "@babel/core": "^7.9.0",
-    "babel": "^6.23.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^8.1.0",
-    "babel-preset-es2015": "^6.24.1",
-    "chokidar": "^3.4.3",
-    "esbuild": "^0.8.26",
-    "lodash": "^4.17.20",
+    "esbuild-loader": "^2.13.1",
     "rimraf": "^3.0.2",
-    "serve-handler": "^6.1.3",
-    "ssr-window": "^3.0.0",
-    "tiny-glob": "^0.2.8"
+    "webpack": "^5.47.0",
+    "webpack-cli": "^4.7.2",
+    "webpack-dev-server": "^3.11.2",
+    "webpack-merge": "^5.8.0"
   },
-  "dependencies": {},
-  "babel": {
-    "presets": [
-      "es2015"
-    ]
-  }
+  "dependencies": {}
 }
