+++ webpack.dev.js
@@ -1,0 +1,11 @@
+const { merge } = require('webpack-merge');
+const common = require('./webpack.common.js');
+
+module.exports = merge(common, {
+  devtool: 'inline-source-map',
+  mode: 'development',
+  devServer: {
+    publicPath: '/dist',
+    contentBase: './',
+  },
+});
